+++ spec/integration/resource/api800/c7000/hypervisor_cluster_profile/create_spec.rb
@@ -1,0 +1,8 @@
+require 'spec_helper'
+
+klass = OneviewSDK::API800::C7000::HypervisorClusterProfile
+RSpec.describe klass, integration: true, type: CREATE, sequence: seq(klass) do
+  include_examples 'HypervisorClusterProfileCreateExample', 'integration api800 context' do
+    let(:current_client) { $client_800 }
+  end
+end