+++ spec/integration/resource/api800/synergy/hypervisor_cluster_profile/delete_spec.rb
@@ -1,0 +1,8 @@
+require 'spec_helper'
+
+klass = OneviewSDK::API800::Synergy::HypervisorClusterProfile
+RSpec.describe klass, integration: true, type: DELETE, sequence: rseq(klass) do
+  include_examples 'HypervisorClusterProfileDeleteExample', 'integration api800 context' do
+    let(:current_client) { $client_800 }
+  end
+end